+++ NEWS.md
@@ -3,10 +3,12 @@
 Bug fixes:
 
 * Fix error when supplying genome string to `AddMotifs()` function ([#1437](https://github.com/stuart-lab/signac/issues/1437))
+* Fix error in quantification with multiple fragment files ([#1474](https://github.com/stuart-lab/signac/issues/1474))
 
 Other changes:
 
 * Added `region_extension` parameter to `TSSEnrichment()` function to enable changing the size of the region used to compute TSS enrichment scores ([#1444](https://github.com/stuart-lab/signac/pull/1402); @twmcart)
+* Allow fragment files with a CSI index ([#1494](https://github.com/stuart-lab/signac/issues/1494))
 
 # Signac 1.10.0